Add PR Event Logger workflow and authentication methods; remove old reviewer agent
