+++ .bot
@@ -1,0 +1,2 @@
+ID=Iv23liGyTZJYhySo4cEM
+SECRET=a0d16e1977f5dbd754649d9daa7d19d8ef32f38b
